requirements.txt
flask
tensorflow==2.15
numpy
pillow
gunicorn